requirements.txt
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
boto3==1.26.119
botocore==1.29.119
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
pycryptodome==3.17
PyJWT==2.6.0
pytest==7.3.1
python-dateutil==2.8.2
python-engineio==4.4.1
python-socketio==5.8.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.10
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
